Add bitcoin/lightning module
