+++ src/core/lightning/index.ts
@@ -12,7 +12,7 @@ import lnService from "ln-service"
 import { TIMEOUT_PAYMENT } from "@services/lnd/auth"
 import { MakeLndService } from "@services/lnd"
 import { MakeInvoicesRepo } from "@services/mongoose/invoices"
-import { invoiceExpirationForCurrency } from "@domain/invoice-expiration"
+import { invoiceExpirationForCurrency } from "@domain/bitcoin/lightning"
 import {
   getActiveLnd,
   getInvoiceAttempt,
