+++ src/domain/bitcoin/lightning/errors.ts
@@ -1,0 +1,9 @@
+export class LightningError extends Error {
+  name = this.constructor.name
+}
+
+export class LnInvoiceDecodeError extends LightningError {}
+
+export class LightningServiceError extends LightningError {}
+export class CouldNotDecodeReturnedPaymentRequest extends LightningServiceError {}
+export class UnknownLightningServiceError extends LightningServiceError {}
