+++ src/domain/bitcoin/lightning/index.types.d.ts
@@ -1,0 +1,37 @@
+type LightningError = import("./errors").LightningError
+type LnInvoiceDecodeError = import("./errors").LnInvoiceDecodeError
+type LightningServiceError = import("./errors").LightningServiceError
+
+declare const invoiceExpirationSymbol: unique symbol
+type InvoiceExpiration = Date & { [invoiceExpirationSymbol]: never }
+
+declare const encodedPaymentRequestSymbol: unique symbol
+type EncodedPaymentRequest = string & { [encodedPaymentRequestSymbol]: never }
+
+declare const paymentHashSymbol: unique symbol
+type PaymentHash = string & { [paymentHashSymbol]: never }
+
+declare const paymentSecretSymbol: unique symbol
+type PaymentSecret = string & { [paymentSecretSymbol]: never }
+
+type LnInvoice = {
+  readonly paymentHash: PaymentHash
+  readonly paymentSecret: PaymentSecret
+  readonly paymentRequest: EncodedPaymentRequest
+}
+
+type RegisterInvoiceArgs = {
+  description: string
+  satoshis: Satoshis
+  expiresAt: InvoiceExpiration
+}
+
+type RegisteredInvoice = {
+  invoice: LnInvoice
+}
+
+interface ILightningService {
+  registerInvoice(
+    registerInvoiceArgs: RegisterInvoiceArgs,
+  ): Promise<RegisteredInvoice | LightningServiceError>
+}
