+++ src/domain/errors.ts
@@ -4,12 +4,6 @@ class CoreError extends Error {
 
 export class AuthorizationError extends CoreError {}
 
-export class LnInvoiceDecodeError extends CoreError {}
-
-export class LightningServiceError extends CoreError {}
-export class CouldNotDecodeReturnedPaymentRequest extends LightningServiceError {}
-export class UnknownLightningServiceError extends LightningServiceError {}
-
 export class RepositoryError extends CoreError {}
 export class UnknownRepositoryError extends RepositoryError {}
 export class CouldNotFindError extends RepositoryError {}
