+++ src/domain/errors.types.d.ts
@@ -1,4 +1,1 @@
-type LnInvoiceDecodeError = import("./errors").LnInvoiceDecodeError
-type LookupError = import("./errors").LookupError
-type LightningServiceError = import("./errors").LightningServiceError
 type RepositoryError = import("./errors").RepositoryError
