--- src/domain/invoice-expiration.types.d.ts
@@ -1,2 +1,0 @@
-declare const invoiceExpirationSymbol: unique symbol
-type InvoiceExpiration = Date & { [invoiceExpirationSymbol]: never }
