--- src/domain/lightning-service.types.d.ts
@@ -1,15 +1,0 @@
-type RegisterInvoiceArgs = {
-  description: string
-  satoshis: Satoshis
-  expiresAt: InvoiceExpiration
-}
-
-type RegisteredInvoice = {
-  invoice: LnInvoice
-}
-
-interface ILightningService {
-  registerInvoice(
-    registerInvoiceArgs: RegisterInvoiceArgs,
-  ): Promise<RegisteredInvoice | LightningServiceError>
-}
