--- src/domain/ln-invoice.types.d.ts
@@ -1,14 +1,0 @@
-declare const encodedPaymentRequestSymbol: unique symbol
-type EncodedPaymentRequest = string & { [encodedPaymentRequestSymbol]: never }
-
-declare const paymentHashSymbol: unique symbol
-type PaymentHash = string & { [paymentHashSymbol]: never }
-
-declare const paymentSecretSymbol: unique symbol
-type PaymentSecret = string & { [paymentSecretSymbol]: never }
-
-type LnInvoice = {
-  readonly paymentHash: PaymentHash
-  readonly paymentSecret: PaymentSecret
-  readonly paymentRequest: EncodedPaymentRequest
-}
